requirements.txt
matplotlib
inquirer
folium
geopandas
pyyaml
